Add support for super breaks and broken loops
The previous version appeared to have support for super breaks (it asked
all the relevant questions during setup), but there actually wasn't
support at run-time. Oops.

Also, this version adds the ability to break a loop and start it again
later. Oh, and it adds some nice debug hooks! Additionally, change the
test script to no longer exit 1 so that `np` can be used to publish.